resolve pr comments
